requirements.txt
Flask-HTTPAuth==4.4.0
PyJWT==1.7.1
Werkzeug==2.3.7